requirements.txt
Django==2.0.6
django-polymorphic==2.0.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.8.1
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.4
